% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate variants from a vector of variant IDs}
\usage{
annotate_variants(rsid, verbose = FALSE)
}
\arguments{
\item{rsid}{a vector of variant IDs}

\item{verbose}{if TRUE will print progress messages (default = FALSE)}
}
\value{
a data.frame with the following columns:
\itemize{
  \item rsid (variant id)
  \item cadd phred (phred score from CADD, higher more deleterious)
  \item fathmm_xf_score (from 0 to 1, higher is more deleterious)
  \item fathmm_xf_pred ("D"(DAMAGING) if score > 0.5, "N"(NEUTRAL) otherwise)
  \item annot_type (the annotation type from CADD, eg: "Intergenic")
  \item consequence (the consequence from CADD, eg: "DOWNSTREAM")
  \item clinical_significance (risk factor from clinvar, eg: "Benign", "risk factor")
  \item snpeff_ann (impact annotation from snpeff eg: "MODIFIER")
}
}
\description{
use \code{\link[myvariant]{getVariants}} and process the output
to return a data.frame instead of nested lists.
}
\examples{
annotate_variants(rsid = c("rs1225680362", "rs12395043", "rs746318172"))
}
