% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_to_gene_ensembl}
\alias{connect_to_gene_ensembl}
\title{Connect to gene Mart}
\usage{
connect_to_gene_ensembl(mirror = "www")
}
\arguments{
\item{mirror}{(optional) to use an alternative mirror. see
\code{\link[biomaRt]{useEnsembl}} for the possible values (default:"www")}
}
\value{
a Mart object with a connection to the "hsapiens_gene_ensembl" dataset
}
\description{
Connect to the "hsapiens_gene_ensembl" dataset in the "ENSEMBL_MART_ENSEMBL"
BioMart database, using \code{\link[biomaRt]{useEnsembl}}
}
\examples{
# Connect with the default mirror
gene_mart <- connect_to_gene_ensembl()

# Connect with the "www" mirror
gene_mart <- connect_to_gene_ensembl(mirror="www")
}
