% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_to_snp_ensembl}
\alias{connect_to_snp_ensembl}
\title{Connect to snp Mart}
\usage{
connect_to_snp_ensembl(mirror = "www")
}
\arguments{
\item{mirror}{(optional) to use an alternative mirror. see
\code{\link[biomaRt]{useEnsembl}} for the possible values (default:"www")}
}
\value{
a Mart object with a connection to the "hsapiens_snp" dataset
}
\description{
Connect to the "hsapiens_snp" dataset in the "snp" BioMart
database, using \code{\link[biomaRt]{useEnsembl}}
}
\examples{
# Connect with the default mirror
ensembl <- connect_to_snp_ensembl()

# Connect with the "www" mirror
www_ensembl <- connect_to_snp_ensembl(mirror="www")
}
