% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{convert_gtex_id_to_granges}
\alias{convert_gtex_id_to_granges}
\title{Convert GTEx IDs to GRanges}
\usage{
convert_gtex_id_to_granges(gtex_ids, hg19ToHg38.over.chain,
  verbose = FALSE)
}
\arguments{
\item{gtex_ids}{a vector of gtex_ids in the format "chr_pos_ref_alt_build"}

\item{hg19ToHg38.over.chain}{the chain file to liftOver locations from
hg19 to hg38. GTEx reports positions based on hg19. Only needed for GTEx
versions < 7}

\item{verbose}{if true, will print information about the conversion (default: FALSE)}
}
\value{
a GRanges object, with the variant locations on hg38.
}
\description{
GTEx IDs are in the format "chr_pos_ref_alt_build".
Transform the IDs to a GRanges object and get the rsids based on the position.
If the build is equals to "b37", then liftOver is performed.
If the GTEx variant is an InDel (ref or alt) then we do position+1 to make it
compatible with ensembl positions.
}
