% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{convert_gtex_to_rsids}
\alias{convert_gtex_to_rsids}
\title{Convert GTEx IDs to rsid}
\usage{
convert_gtex_to_rsids(gtex_ids, gtex_lookup, verbose = FALSE)
}
\arguments{
\item{gtex_ids}{a vector of gtex_ids in the format "chr_pos_ref_alt_build"}

\item{gtex_lookup}{a data.frame representing the lookup file. It should contains
three columns: the gtex ids v8, the rsids, the gtex ids v7. The columns should
be respectively called "variant_id", "rs_id_dbSNP151_GRCh38p7" and "variant_id_b37".
This can be created by using fread() on the lookup file (see get_gtex_variants)}

\item{verbose}{if true, will print information about the conversion (default: FALSE)}
}
\value{
a vector of rsids (as some gtex ids do not have a corresponding rsid
the output can be smalled than the input)
}
\description{
GTEx IDs are in the format "chr_pos_ref_alt_build". To make it
consistent with the rest of the package, we are converting them to rsids.
We us the "gtex_lookup" which is a data.frame created by reading the gtex lookup file.
(see: get_gtex_variants)
}
