% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_chr}
\alias{format_chr}
\title{Format chromosome names}
\usage{
format_chr(chr)
}
\arguments{
\item{chr}{a vector of chromosome names}
}
\value{
the formatted vector of chromosome names (eg: "chr1" instead of "1")
}
\description{
Transform 1 to 22, X, Y and MT into chr1 to chr22, chrX, chrY and chMT.
Regex needed because some chr are named CHR17_... and we do not want chrCHR17
}
