% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_output}
\alias{format_output}
\title{Format the variants data.frame}
\usage{
format_output(chr, pos, rsid, ensembl_gene_id, hgnc_symbol)
}
\arguments{
\item{chr}{list of variant chromosome}

\item{pos}{list of variant positions}

\item{rsid}{list of variant rsid}

\item{ensembl_gene_id}{gene id associated to the variants}

\item{hgnc_symbol}{hgnc_symbol associated to the variants}
}
\value{
a data.frame of variants with the following columns
\itemize{
  \item chr (chromosome)
  \item pos (position of the variant)
  \item rsid (variant ID)
  \item ensembl_gene_id ("gene id" of the gene associated with the variants)
  \item hgnc_symbol ("hgnc symbol" of the gene associated with the variants)
}
}
\description{
Used to have consistent output for the variants from OMIM, FANTOM5,
GTEx and GWAS.
}
