% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantom.R
\name{get_fantom5_enhancers_from_hgnc}
\alias{get_fantom5_enhancers_from_hgnc}
\title{Get the enhancers associated to certain genes from FANTOM5}
\usage{
get_fantom5_enhancers_from_hgnc(fantom_df, hgnc_symbols, corr_threshold = 0.25)
}
\arguments{
\item{fantom_df}{the output of \code{\link{prepare_fantom}}}

\item{hgnc_symbols}{vector of HUGO ids for the genes of interest}

\item{corr_threshold}{the minimum correlation (z-score) to consider a
enhancer/gene assocation valid (default: 0.25).
A z-score greater than 0 represents an element greater than the mean, this means
that this association has more correlation than random motifs.}
}
\value{
a subset of the FANTOM5 data.frame containing the information about the
enhancers. The data.frame contains the following columns
\itemize{
  \item chr (chromosome)
  \item start (start of the enhancers)
  \item end (end of the enhancers)
  \item symbol (HGNC symbol of the gene associated to the enhancer)
  \item corr (correlation z-score from FANTOM5)
  \item fdr (False Discovery Rate)
}
}
\description{
from the FANTOM5 dataset, get the enhancers associated to the
genes using the HGNC symbols, the association must pass the correlation threshold given
by the user. Is used internally by \code{\link{get_fantom5_variants}}
}
