% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantom.R
\name{get_fantom5_variants}
\alias{get_fantom5_variants}
\title{Get variants on the enhancers of the list of genes given as input}
\usage{
get_fantom5_variants(
  fantom_df,
  omim_genes,
  corr_threshold = 0.25,
  hg19ToHg38.over.chain,
  verbose = FALSE
)
}
\arguments{
\item{fantom_df}{the output of \code{\link{prepare_fantom}}}

\item{omim_genes}{output from \code{\link{get_omim_genes}}}

\item{corr_threshold}{the minimum correlation (z-score) to consider a
enhancer/gene association valid (default: 0.25). A z-score greater than 0
represents an element greater than the mean, this means that this association
 has more correlation than random motifs.}

\item{hg19ToHg38.over.chain}{the chain file to liftOver locations from
hg19 to hg38.}

\item{verbose}{if true, will print progress information (default: FALSE)}
}
\value{
a data.frame with information about the variants located on the enhancers.
The data.frame will contain the following columns:
\itemize{
  \item chr (chromosome)
  \item pos (position of the variant)
  \item rsid (variant ID)
  \item ensembl_gene_id ("gene id" of the gene associated with the variant)
  \item hgnc_symbol ("hgnc symbol" of the gene associated with the variant)
  \item source (here the value will be "fantom5")
}
}
\description{
FANTOM5 is used to get the enhancers of the genes, then the variants
located on the enhancers are fetched with \code{\link{get_variants_from_locations}}
It is also possible to specify a correlation threshold to limit the number
of association to output (default: 0.25). The enhancer tss association file
is based on hg19, so liftOver is needed.
}
