% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omim.R
\name{get_genes_variants}
\alias{get_genes_variants}
\title{Get the variants located on the genes}
\usage{
get_genes_variants(genes, verbose = FALSE)
}
\arguments{
\item{genes}{list of genes (can be obtained from \code{\link{get_omim_genes}})}

\item{verbose}{if true, will print progress information (default: FALSE)}
}
\value{
a data.frame of variants with the following columns (see: \code{\link{format_output}})
\itemize{
  \item chr (chromosome)
  \item pos (position of the variant)
  \item rsid (variant ID)
  \item ensembl_gene_id ("gene id" of the gene associated with the variant)
  \item hgnc_symbol ("hgnc symbol" of the gene associated with the variant)
  \item source (here the value will be "omim")
}
}
\description{
Get the variants located between the start and stop positions
of the genes given as input.
}
