% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omim.R
\name{get_omim_genes}
\alias{get_omim_genes}
\title{Get ensembl gene IDs from "mim_morbid_accession"}
\usage{
get_omim_genes(omim_ids, gene_mart)
}
\arguments{
\item{omim_ids}{is a vector of morbity accession
eg: c("125853","222100") for Diabetes}

\item{gene_mart}{a connection to hsapiens gene mart, can be generated
from \code{\link{connect_to_gene_ensembl}} (which will be used if this argument
if not specified).}
}
\value{
a dataframe with
\itemize{
 \item ensembl_gene_id (ensembl stable gene ID, eg:"ENSG00000134640")
 \item chromosome_name (chromosome name)
 \item start_position (gene start)
 \item end_position (gene end)
 \item hgnc_symbol (gene id from the Hugo Gene Nomenclature Committee, eg:"MTNR1B")
 \item mim_morbid_accession (the OMIM morbid ID, eg:"125853")
 \item mim_morbid_accession (the OMIM description, eg: "NONINSULIN-DEPENDENT DIABETES MELLITUS")
}
}
\description{
Use \code{\link[biomaRt]{getBM}} to filter results by
"mim_morbid_accession". One or more accessions can be given as input.
}
\examples{
gene_mart <- connect_to_gene_ensembl()
DM_genes <- get_omim_genes(c("125853","222100"), gene_mart)
}
