% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{get_omim_variants}
\alias{get_omim_variants}
\title{Get the variants on the OMIM genes}
\usage{
get_omim_variants(omim_genes, verbose = FALSE)
}
\arguments{
\item{omim_genes}{list of omim genes from \code{\link{get_omim_genes}}}

\item{verbose}{if true, will print progress information (default: FALSE)}
}
\value{
a data.frame of variants with the following columns (see: \code{\link{format_output}})
\itemize{
  \item chr (chromosome)
  \item pos (position of the variant)
  \item rsid (variant ID)
  \item ensembl_gene_id ("gene id" of the gene associated with the variant)
  \item hgnc_symbol ("hgnc symbol" of the gene associated with the variant)
  \item source (here the value will be "omim")
}
}
\description{
Get the variants on the gene locations obtained from
\code{\link{get_omim_genes}}
}
