% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{get_phenotype_terms}
\alias{get_phenotype_terms}
\title{From a keyword, get a list of phenotypes to use with snp mart}
\usage{
get_phenotype_terms(keyword, snp_mart)
}
\arguments{
\item{keyword}{a keyword, all the phenotypes contaning this keyword will be
returned (based on \code{\link[base]{grep}} with ignore.case)}

\item{snp_mart}{a connection to ensembl snp mart, can be generated from
\code{\link{connect_to_snp_ensembl}}}
}
\value{
a vector of phenotype names
}
\description{
The list of phenotypes obtained can be used with
\code{\link{get_variants_from_phenotypes}} to get a list of variants associated
with the phenotypes.
}
\examples{
# First, connect to snp ensembl
snp_mart <- connect_to_snp_ensembl()

# Get the phenotypes related to diabetes
get_phenotype_terms(keyword = "diabetes", snp_mart = snp_mart)
}
