% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_variants_from_locations}
\alias{get_variants_from_locations}
\title{Using ensembl API to get the variants at certain locations}
\usage{
get_variants_from_locations(locations, verbose = FALSE)
}
\arguments{
\item{locations}{a vector of locations (eg: c("chr1:1000:10000", "chr2:10-100"))}

\item{verbose}{if TRUE, will print progress messages (default: FALSE)}
}
\value{
a data.frame containing the following columns:
\itemize{
  \item clinical_significance
  \item start
  \item consequence_type
  \item alleles
  \item seq_region_name
  \item feature_type
  \item source
  \item strand
  \item assembly_name
  \item end
  \item id
}
}
\description{
Use \code{\link[httr]{GET}} to query the ensembl REST API to
gather variants located on specific locations. The content is fetched as a
JSON and then translated as a data.frame, using \code{\link[jsonlite]{toJSON}}
/!\ the column order is sometimes different.
}
\examples{
get_variants_from_locations(c("chr1:100:1000", "chr10:2004:12042"))
}
