% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varphen.R
\name{get_variants_from_phenotypes}
\alias{get_variants_from_phenotypes}
\title{Get variants associated with certain phenotypes}
\usage{
get_variants_from_phenotypes(phenotypes, snp_mart)
}
\arguments{
\item{phenotypes}{a vector of phenotypes of interest (obtained from
\code{\link{get_phenotype_terms}})}

\item{snp_mart}{a connection to ensembl snp mart, can be generated
from \code{\link{connect_to_snp_ensembl}} (which will be used if this argument
is not provided)}
}
\value{
a data.frame of variants associated with the phenotypes:
\itemize{
  \item chr_name (chromosome)
  \item chrom_start (variant position)
  \item refsnp_id (variant id)
  \item refsnp_source (the source of information (eg: "dbSNP"))
  \item ensembl_transcript_chrom_strand (strand information (1, -1 or NA))
  \item associated_variant_risk_allele (risk allele (eg: "A"))
  \item phenotype_description (description of the phenotype (eg: "Diabetes mellitus type 1"))
  \item clinical_significance (from clinVar (eg: "pathogenic", "begnin" ...))
  \item validated (Variant supporting evidence (1000Genomes, TOPMed, etc...))
}
}
\description{
The vector of phenotypes can be obtained with
\code{\link{get_phenotype_terms}}
}
\examples{
# First, connect to ensembl snp mart
snp_mart <- connect_to_snp_ensembl()

# Get the phenotypes in biomaRt containing the word "Diabetes"
DM_phen <- get_phenotype_terms(keywords = "diabetes", snp_mart = snp_mart)

# Get all the variants associated with the penotypes in "DM_phen"
get_variants_from_phenotypes(phenotypes = DM_phen, snp_mart = snp_mart)
}
