% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex.R
\name{list_gtex_tissues}
\alias{list_gtex_tissues}
\title{Generates a vector containing the available GTEx tissue files}
\usage{
list_gtex_tissues(gtex_dir)
}
\arguments{
\item{gtex_dir}{is the directory contaning the files from
"GTEx_Analysis_v8_eQTL.tar.gz" (cf: https://gtexportal.org/home/datasets)}
}
\value{
data.frame containing the full file path of each tissue and keywords to select them
}
\description{
The directory can be downloaded with:
download.file("http://www.ebi.ac.uk/gwas/api/search/downloads/alternative",
}
\examples{
vargen_install(install_dir = "./vargen_data/")
list_gtex_tissues("./vargen_data/GTEx_Analysis_v8_eQTL/")
}
