% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{list_gwas_traits}
\alias{list_gwas_traits}
\title{List the available gwas traits}
\usage{
list_gwas_traits(keywords = "", gwascat_file)
}
\arguments{
\item{keywords}{a vector of keywords to grep the traits from the gwas catalog. (default: "")}

\item{gwascat_file}{(optional) a path to the file containing a gwas_catalog
(available at http://www.ebi.ac.uk/gwas/api/search/downloads/alternative). The
function will use the filename to determine the extract date, please have it
in the format: \[filename\]_r**YYYYY**-**MM**-**DD**.tsv, eg:
"gwas_catalog_v1.0.2-associations_e96_r2019-07-30.tsv"}
}
\value{
a vector of traits
}
\description{
Return the gwas traits available in the gwas catalog, based on a keyword.
The traits are found using \code{\link[base]{grep}} on the `DISEASE/TRAIT`
column from the gwas object produced by \code{\link[gwascat]{makeCurrentGwascat}}.
Output can be used as parameter for \code{\link{vargen_pipeline}}
}
\examples{
list_gwas_traits(keywords = c("type 1 diabetes", "Obesity"))
}
