% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{list_omim_accessions}
\alias{list_omim_accessions}
\title{List the available OMIM morbid IDs and descriptions}
\usage{
list_omim_accessions(gene_mart)
}
\arguments{
\item{gene_mart}{a connection to hsapiens gene mart, can be generated
from \code{\link{connect_to_gene_ensembl}} (which will be used if this argument
if not specified).}
}
\value{
a dataframe with two columns ("mim_morbid_accession" and "mim_morbid_description")
}
\description{
List the "OMIM IDs" available in biomaRt, will also
print the corresponding description. Use \code{\link[biomaRt]{getBM}} with the
following filter: "with_mim_morbid".
This function is useful to get OMIM IDs for \code{\link{get_omim_genes}}
}
\examples{
gene_mart <- connect_to_gene_ensembl()
list_omim_accessions(gene_mart)
}
