% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{plot_manhattan_gwas}
\alias{plot_manhattan_gwas}
\title{Manhattan plot for variants found in GWAS}
\usage{
plot_manhattan_gwas(traits, gwas_cat, list_chr)
}
\arguments{
\item{traits}{a vector with the trait of interest (as characters). The list
of available traits can be obtained with \code{\link{list_gwas_traits}}}

\item{gwas_cat}{a gwaswloc object obtained from \code{\link{create_gwas}}}

\item{list_chr}{(optional) a list of chromosome to plot. (format: "chr{1-22-X-Y})}
}
\value{
nothing (just display the plot)
}
\description{
Display a manhattan plot. Only the variants related to the traits
given as parameter will be displayed. The two horizontal lines on the plot
correspond to the "suggestive" and "significant" thresholds in genome wide
studies.
}
\examples{
gwas_cat <- create_gwas()

plot_manhattan_gwas(gwas_cat = gwas_cat, traits = c("Type 1 diabetes", "Type 2 diabetes"))
}
\references{
Lander E, Kruglyak L. Genetic dissection of complex traits: guidelines for
interpreting and reporting linkage results. Nat Genet. 1995;11(3):241-7.
}
