% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantom.R
\name{prepare_fantom}
\alias{prepare_fantom}
\title{Generate data.frame from FANTOM5 enhancer file}
\usage{
prepare_fantom(enhancer_tss_association)
}
\arguments{
\item{enhancer_tss_association}{the "enhancer_tss_associations.bed" file from FANTOM5.
The file can be downloaded here:
http://enhancer.binf.ku.dk/presets/enhancer_tss_associations.bed}
}
\value{
a data.frame containing the file information
}
\description{
Prepare FANTOM5 for \code{\link{get_fantom5_enhancers_from_hgnc}}
}
