% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex.R
\name{select_gtex_tissues}
\alias{select_gtex_tissues}
\title{Get the GTEx files corresponding to the tissues of interest}
\usage{
select_gtex_tissues(gtex_dir, tissues_query = "")
}
\arguments{
\item{gtex_dir}{is the directory contaning the files from
"GTEx_Analysis_v8_eQTL.tar.gz" (cf: https://gtexportal.org/home/datasets)}

\item{tissues_query}{is a vector of keywords to select the tissues of interest
(eg: "adipose", "brain", "Heart_Left_Ventricle"). NOT case sensitive. (default:
"", return everything)}
}
\value{
the list of files corresponding to the tissues of interest
}
\description{
Use \code{\link{list_gtex_tissues}} to get the list of GTEx files
and their associated keywords. Grep the user's query against the list of
keywords (case is ignored) to return all the corresponding tissue files.
eg: "adipose" will return the files corresponding to "Adipose_Subcutaneous" AND
"Adipose_Visceral_Omentum". The tissues in the query are grepped one after the
other.
If one of the keywords does not correspond to any tissues, \code{\link[base]{stop}}
is called and an error message is displayed.
}
\examples{
vargen_install(install_dir = "./vargen_data/")
select_gtex_tissues("./vargen_data/GTEx_Analysis_v8_eQTL/",
                    c("adipose", "Heart_Left_Ventricle", "leg"))

select_gtex_tissues("./vargen_data/GTEx_Analysis_v8_eQTL/",
                    "brain")
}
