% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vargen.R
\name{vargen_visualisation}
\alias{vargen_visualisation}
\title{Generate a plot of the variants on the OMIM genes}
\usage{
vargen_visualisation(
  annotated_snps,
  outdir = "./",
  rsid_highlight,
  device = "pdf",
  verbose = FALSE,
  gene_mart
)
}
\arguments{
\item{annotated_snps}{a data.frame from the merging of the outputs of
\code{\link{vargen_pipeline}} and \code{\link{annotate_variants}}}

\item{outdir}{the directory that will contain the plots}

\item{rsid_highlight}{optional, a vector of rsids, which will be plotted in red
on the cadd track.}

\item{device}{only "pdf" and "png" are supported now.}

\item{verbose}{if TRUE, will display progress messages}

\item{gene_mart}{optional, a connection to ensembl gene mart, can be created
using \code{\link{connect_to_gene_ensembl}}. If missing this function will be
used to create the connection.}
}
\value{
nothing, create the plots in "outdir". One file per gene, the name format
is <hgnc_symbol>_<ensembl_id>_GVIZ
}
\description{
The plot contains 4 tracks:
\itemize{
  \item  The chromosome, with a red marker on the gene location
  \item  The ensembl transcripts
  \item  The variant consequences, grouped by type (eg: INTRONIC, STOP LOST etc...),
each green bar represent a variant
  \item  The cadd phred score, each dot represent a variant
}
}
\examples{
vargen_install("./vargen_data/")

# Simple query
gene_mart <- connect_to_gene_ensembl()
DM1_simple <- vargen_pipeline(vargen_dir = "./vargen_data/", omim_morbid_ids = "222100",
                              fantom_corr = 0.25, outdir = "./", verbose = TRUE)

vargen_visualisation(annotated_snps = DM1_simple, verbose = TRUE,
                     outdir = "./DM1_gviz", device = "png",
                     gene_mart = gene_mart)
}
