% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{convert_gtex_to_rsids}
\alias{convert_gtex_to_rsids}
\title{Convert GTEx IDs to rsid}
\usage{
convert_gtex_to_rsids(gtex_ids, hg19ToHg38.over.chain, verbose = FALSE)
}
\arguments{
\item{gtex_ids}{a vector of gtex_ids in the format "chr_pos_ref_alt_build"}

\item{hg19ToHg38.over.chain}{the chain file to liftOver locations from
hg19 to hg38. GTEx reports positions based on hg19.}

\item{verbose}{if true, will print information about the conversion (default: FALSE)}
}
\value{
the same as \code{\link{get_variants_from_locations}}
}
\description{
GTEx IDs are in the format "chr_pos_ref_alt_build".
Transform the IDs to a dataframe and get the rsids based on the position.
If the build is equals to "b37", then liftOver is performed.
Use \code{\link{convert_gtex_id_to_granges}}
}
\examples{
vargen_install(install_dir = "./vargen_data/")

convert_gtex_to_rsids(c("1_98929_A_G_b37", "10_61803062_T_G_b37"),
                        "./vargen_data/hg19ToHg38.over.chain", FALSE)

}
