% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex.R
\name{convert_gtex_to_rsids}
\alias{convert_gtex_to_rsids}
\title{Convert GTEx IDs to rsid}
\usage{
convert_gtex_to_rsids(gtex_variants, gtex_lookup_file, verbose = FALSE)
}
\arguments{
\item{gtex_variants}{a data.frame of gtex variants obtained during "get_gtex_variants"
should contain at least one column with the gtex_id as "chr_pos_alt_ref_build"}

\item{gtex_lookup_file}{the lookup file, GTEx to rsids. "GTEx_Analysis_2017-06-05_v8_WholeGenomeSeq_838Indiv_Analysis_Freeze.lookup_table.txt.gz"
Can be obtained using \code{\link{vargen_install}}.}

\item{verbose}{if true, will print information about the conversion (default: FALSE)}
}
\value{
a vector of rsids (as some gtex ids do not have a corresponding rsid
the output can be smalled than the input)
}
\description{
GTEx IDs are in the format "chr_pos_ref_alt_build". To make it
consistent with the rest of the package, we are converting them to rsids.
We us the "gtex_lookup" which is a data.frame created by reading the gtex lookup file.
(see: get_gtex_variants)
}
