% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{create_gwas}
\alias{create_gwas}
\title{Create a gwaswloc object}
\usage{
create_gwas(gwascat_file)
}
\arguments{
\item{gwascat_file}{(optional) a path to the file containing a gwas_catalog
(available at http://www.ebi.ac.uk/gwas/api/search/downloads/alternative).}
}
\description{
Create a gwaswloc object from \code{\link[gwascat]{makeCurrentGwascat}}
or by reading the "gwascat_file" (if present as a parameter).
The function will use the filename to determine the extract date, please have
it in the format: \[filename\]_r**YYYYY**-**MM**-**DD**.tsv
eg: "gwas_catalog_v1.0.2-associations_e96_r2019-07-30.tsv"
}
\examples{
gwas_cat <- create_gwas()
}
