% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex.R
\name{get_gtex_variants}
\alias{get_gtex_variants}
\title{Get variants from GTEx linked to the given ensembl genes}
\usage{
get_gtex_variants(
  tissue_files,
  omim_genes,
  gtex_lookup_file,
  snp_mart,
  verbose = FALSE
)
}
\arguments{
\item{tissue_files}{a vector containing the name of the "signif_variant_gene_pairs.txt.gz"
files. This will be read using \code{\link[base]{gzfile}}. Output from
\code{\link{select_gtex_tissues}} can be used.}

\item{omim_genes}{output from \code{\link{get_omim_genes}}}

\item{gtex_lookup_file}{the lookup file, GTEx to rsids. "GTEx_Analysis_2017-06-05_v8_WholeGenomeSeq_838Indiv_Analysis_Freeze.lookup_table.txt.gz"
Can be obtained using \code{\link{vargen_install}}.}

\item{snp_mart}{optional, a connection to ensembl snp mart, can be created
using \code{\link{connect_to_snp_ensembl}} (If missing this function will be
used to create the connection).}

\item{verbose}{will be given to subsequent functions to print progress.}
}
\value{
a data.frame with information about the variants associated with a
change in expression on the gene of interest.
The data.frame will contain the following columns:
\itemize{
  \item chr (chromosome)
  \item pos (position of the variant)
  \item rsid (variant ID)
  \item ensembl_gene_id ("gene id" of the gene associated with the variant)
  \item hgnc_symbol ("hgnc symbol" of the gene associated with the variant)
  \item source (here the value will be "gtex")
}
}
\description{
Take as input one or more tissue files from
 "GTEx_Analysis_v8_eQTL" (or v7), as well as a vector of ensembl gene ids.
 This function will return the variants that are associated with changes
 in the expression of the selected genes in the selected tissues. The
 assocations are based on the "signif_variant_gene_pairs" files.
 The ensembl ids from the GTEx file will be converted to stable ids.
}
