% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varphen.R
\name{get_phenotype_terms}
\alias{get_phenotype_terms}
\title{From keywords, get a list of phenotypes to use with snp mart}
\usage{
get_phenotype_terms(keywords, snp_mart)
}
\arguments{
\item{keywords}{a keyword, all the phenotypes contaning this keyword will be
returned (based on \code{\link[base]{grep}} with ignore.case)}

\item{snp_mart}{optional, a connection to ensembl snp mart, can be created
using \code{\link{connect_to_snp_ensembl}} (If missing this function will be
used to create the connection).}
}
\value{
a vector of phenotype names
}
\description{
The list of phenotypes obtained can be used with
\code{\link{get_variants_from_phenotypes}} to get a list of variants associated
with the phenotypes.
}
\examples{
# First, connect to snp ensembl
snp_mart <- connect_to_snp_ensembl()

# Get the phenotypes related to diabetes
get_phenotype_terms(keywords = "diabetes", snp_mart = snp_mart)
}
