% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGen.R
\name{list_gwas_traits}
\alias{list_gwas_traits}
\title{List the available gwas traits}
\usage{
list_gwas_traits(keywords = "", vargen_dir)
}
\arguments{
\item{keywords}{a vector of keywords to grep the traits from the gwas catalog. (default: "")}

\item{vargen_dir}{(optional) a path to vargen data directory, created 
during \code{\link{vargen_install}}. If not specified, the gwas object will
be created with \code{\link[gwascat]{makeCurrentGwascat}}. If specified, this
function will look for files that begin with "gwas_catalog". If more than one 
is found, the user will have to choose one via a text menu}
}
\value{
a vector of traits
}
\description{
Return the gwas traits available in the gwas catalog, based on keywords.
The traits are found using \code{\link[base]{grep}} on the `DISEASE/TRAIT`
column from the gwas object produced by \code{\link[gwascat]{makeCurrentGwascat}}.
Output can be used as parameter for \code{\link{vargen_pipeline}}
}
\examples{
list_gwas_traits(keywords = c("type 1 diabetes", "Obesity"))
}
