% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{list_gwas_traits}
\alias{list_gwas_traits}
\title{List the available gwas traits}
\usage{
list_gwas_traits(keywords = "", gwas_cat)
}
\arguments{
\item{keywords}{a vector of keywords to grep the traits from the gwas catalog. (default: "")}

\item{gwas_cat}{output from \code{\link{create_gwas}}}
}
\value{
a vector of traits
}
\description{
Return the gwas traits available in the gwas catalog, based on keywords.
The traits are found using \code{\link[base]{grep}} on the `DISEASE/TRAIT`
column from the gwas object produced by \code{\link{create_gwas}}.
Output can be used as parameter for \code{\link{vargen_pipeline}}
}
\examples{
list_gwas_traits(keywords = c("type 1 diabetes", "Obesity"))
}
